+++ philo_one/srcs/utils/time_start.c
@@ -17,6 +17,6 @@ ssize_t 	time_start(void)
 	ssize_t	s_time;
 	struct timeval	t_start;
 	gettimeofday(&t_start, NULL);
-	s_time = (t_start.tv_sec * 1000 + t_start.tv_usec * 0.001);
+	s_time = (ssize_t)((double)t_start.tv_sec * 1000 + t_start.tv_usec * 0.001);
 	return (s_time);
 }
